SolaraAIQuant/__OLD/requirements.txt
# ─────────────────────────────────────────────
# Solara AI Quant — Production Dependencies
# ─────────────────────────────────────────────

# MT5 connectivity
MetaTrader5>=5.0.45

# Data
pandas>=2.1.0
numpy>=1.26.0

# ML inference
scikit-learn>=1.4.0
joblib>=1.3.0

# Config & validation
pydantic>=2.5.0
pyyaml>=6.0.1

# File watchdog
watchdog>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Logging
structlog>=24.1.0

# Environment
python-dotenv>=1.0.0
